requirements.txt
lime==0.2.0.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
plotly==5.3.1
streamlit==0.89.0
seaborn==0.11.1
lightgbm==3.2.1
scikit-learn==0.24.1
st-annotated-text==2.0.0